+++ devtools/src/webtransport-devtools/components/index.ts
@@ -1,6 +1,7 @@
 export { ActionButton } from "./ActionButton";
 export { MessageItem } from "./MessageItem";
 export { ConnectionPanel } from "./ConnectionPanel";
+export { PostConnectionPanel } from "./PostConnectionPanel";
 export { StaticApiSupportPanel } from "./StaticApiSupportPanel";
 export { BidiStreamPanel } from "./BidiStreamPanel";
 export { UniSendStreamPanel } from "./UniSendStreamPanel";